+++ frontend/app/src/utils/cfmmUtils.ts
@@ -61,3 +61,28 @@ export const tradeDCashForDTez = (
   }
   return dtezApprox;
 };
+
+export const calculateMarginalPrice = (tez: number, cash: number, target: number): number => {
+  const x = cash * target;
+  const y = tez * 2 ** 48;
+  const x2 = x * x;
+  const y2 = y * y;
+  const nom = tez * (3 * x2 + y2);
+  const denom = cash * (3 * y2 + x2);
+  return (nom * 2 ** 48) / denom / 2 ** 48;
+};
+
+export const isMonthFromLiquidation = (
+  outstandingCtez: number,
+  target: number,
+  tezBalance: number,
+  currentDrift: number,
+): boolean => {
+  return (
+    outstandingCtez *
+      (target / 2 ** 48) *
+      (1 + currentDrift / 2 ** 48) ** ((365.25 * 24 * 3600) / 12) *
+      (16 / 15) >
+    tezBalance
+  );
+};
